refactor: avoid cod duplication on literal destructuring
