+++ src/rules/classnames-order.ts
@@ -13,8 +13,8 @@ import {
   PluginSettings,
 } from "../utils/parse-plugin-settings";
 import {
+  dissectAtomicNode,
   getClassnamesFromValue,
-  getTemplateElementAffixes,
 } from "../utils/parser/node";
 import { defineVisitors, GenericRuleContext } from "../utils/parser/visitors";
 import {
@@ -48,91 +48,38 @@ const sortClassnames = (
   literals: Array<AtomicNode>
 ) => {
   for (const node of literals) {
-    let originalClassNamesValue = "";
-    let start = 0;
-    let end = 0;
-    let prefix = "";
-    let suffix = "";
-    switch (node.type) {
-      case TSESTree.AST_NODE_TYPES.Literal: {
-        originalClassNamesValue = "" + node.value;
-        [start, end] = node.range;
-        start++;
-        end--;
-        break;
-      }
-      case TSESTree.AST_NODE_TYPES.TemplateElement: {
-        originalClassNamesValue = node.value.raw;
-        if (originalClassNamesValue === "") {
-          break;
-        }
-        [start, end] = node.range;
-        // https://github.com/eslint/eslint/issues/13360
-        // The problem is that range computation includes the backticks (`test`)
-        // but `value.raw` does not include them, so there is a mismatch.
-        // start/end does not include the backticks, therefore it matches value.raw.
-        const rawCode = context.sourceCode.getText(
-          node as unknown as TSESTree.Node
-        );
-        [prefix, suffix] = getTemplateElementAffixes(
-          rawCode,
-          originalClassNamesValue
-        );
-        break;
-      }
-      case "TextAttribute": {
-        originalClassNamesValue = node.value;
-        start = node.valueSpan.fullStart.offset;
-        end = node.valueSpan.end.offset;
-        break;
-      }
-      case "VLiteral": {
-        originalClassNamesValue = "" + node.value;
-        [start, end] = node.range;
-        start++;
-        end--;
-        break;
-      }
-      default: {
-        // console.log(index, "Unhandled literal type", literal.type);
-        break;
+    const { originalClassNamesValue, start, end, prefix, suffix } =
+      dissectAtomicNode(node, context as unknown as GenericRuleContext);
+    // Process the extracted classnames and report
+    const { classNames, whitespaces, headSpace, tailSpace } =
+      getClassnamesFromValue(originalClassNamesValue);
+    // Skip empty/Single className
+    if (classNames.length <= 1) continue;
+    const orderedClassNames = getSortedClassNamesWorker(
+      settings.cssConfigPath,
+      classNames
+    );
+
+    // Generates the validated/sorted attribute value
+    let validatedClassNamesValue = "";
+    for (let index = 0; index < orderedClassNames.length; index++) {
+      const w = whitespaces[index] ?? "";
+      const cls = orderedClassNames[index];
+      validatedClassNamesValue += headSpace ? `${w}${cls}` : `${cls}${w}`;
+      if (headSpace && tailSpace && index === orderedClassNames.length - 1) {
+        validatedClassNamesValue += whitespaces.at(-1) ?? "";
       }
     }
-    // Process the extracted classnames and report
-    {
-      const { classNames, whitespaces, headSpace, tailSpace } =
-        getClassnamesFromValue(originalClassNamesValue);
-      // Skip empty/Single className
-      if (classNames.length <= 1) continue;
-      const orderedClassNames = getSortedClassNamesWorker(
-        settings.cssConfigPath,
-        classNames
-      );
 
-      // Generates the validated/sorted attribute value
-      let validatedClassNamesValue = "";
-      for (let index = 0; index < orderedClassNames.length; index++) {
-        const w = whitespaces[index] ?? "";
-        const cls = orderedClassNames[index];
-        validatedClassNamesValue += headSpace ? `${w}${cls}` : `${cls}${w}`;
-        if (headSpace && tailSpace && index === orderedClassNames.length - 1) {
-          validatedClassNamesValue += whitespaces.at(-1) ?? "";
-        }
-      }
-
-      if (originalClassNamesValue !== validatedClassNamesValue) {
-        validatedClassNamesValue = prefix + validatedClassNamesValue + suffix;
-        context.report({
-          node: node as TSESTree.Node,
-          messageId: "fix:sort",
-          fix: function (fixer) {
-            return fixer.replaceTextRange(
-              [start, end],
-              validatedClassNamesValue
-            );
-          },
-        });
-      }
+    if (originalClassNamesValue !== validatedClassNamesValue) {
+      validatedClassNamesValue = prefix + validatedClassNamesValue + suffix;
+      context.report({
+        node: node as TSESTree.Node,
+        messageId: "fix:sort",
+        fix: function (fixer) {
+          return fixer.replaceTextRange([start, end], validatedClassNamesValue);
+        },
+      });
     }
   }
 };
